chore: add a function to duplicate all account role key and value
The reason for duplicating the key and value in our `account_roles`
adding two more keys and values suffixed with `New` is one because we
need to keep the original values un-changed since they are required
for the config to not change. Where the suffix `New` can possibly
change so these will be used in making a more user friendly profile name.