feat: Add a agent (storing code) http (server to expose a callback)
Add the parts that are specific to `authorization_code`. High level
the difference is authorization needs to spin up a http
server (http://127.0.0.1:<random-port>/oauth/callback) and
expose a callback that will receive the code that can then be
exchanged for a access token.

The code is stored via an Agent so we don't have to pass messages
around. This can just be called from the http server when the callback
is hit.